# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Import/Export management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Client {
  id: number;
  nom: string;
  telephone: string;
  adresse?: string | null;
  dateCreation: string;
  totalDu: number;
  totalPaye: number;
  solde: number;
}

export interface CreateClientRequest {
  nom: string;
  telephone: string;
  adresse?: string | null;
}

export interface Arrivage {
  id: number;
  dateArrivee: string;
  nbColisTotal: number;
  poidsTotal: number;
  coutTotal: number;
  dateCreation: string;
}

export type ColisType = (typeof ColisType)[keyof typeof ColisType];

export const ColisType = {
  individuel: "individuel",
  mixte: "mixte",
} as const;

export interface Colis {
  id: number;
  arrivageId: number;
  codeColisReel: string;
  codeColisComplet: string;
  type: ColisType;
  poids: number;
  montant: number;
  dateCreation: string;
}

export interface ColisProprietaire {
  id: number;
  colisId: number;
  clientId: number;
  clientNom: string;
  clientTelephone: string;
  poids: number;
  montantDu: number;
  montantPaye: number;
  solde: number;
}

export type ColisDetail = Colis & {
  proprietaires: ColisProprietaire[];
  arrivageDate: string;
};

export type ArrivageDetail = Arrivage & {
  colis: ColisDetail[];
};

export interface CreateArrivageRequest {
  dateArrivee: string;
  nbColisTotal: number;
  poidsTotal: number;
  coutTotal: number;
}

export interface CreateColisProprietaireRequest {
  clientId: number;
  poids: number;
  montantDu: number;
  montantPaye: number;
}

export type CreateColisRequestType =
  (typeof CreateColisRequestType)[keyof typeof CreateColisRequestType];

export const CreateColisRequestType = {
  individuel: "individuel",
  mixte: "mixte",
} as const;

export interface CreateColisRequest {
  arrivageId: number;
  codeColisReel: string;
  type: CreateColisRequestType;
  poids: number;
  montant: number;
  proprietaires: CreateColisProprietaireRequest[];
}

export interface Paiement {
  id: number;
  clientId: number;
  colisProprietaireId: number;
  montant: number;
  datePaiement: string;
  mode: string;
}

export type CreatePaiementRequestMode =
  (typeof CreatePaiementRequestMode)[keyof typeof CreatePaiementRequestMode];

export const CreatePaiementRequestMode = {
  especes: "especes",
  virement: "virement",
  cheque: "cheque",
  mobile: "mobile",
} as const;

export interface CreatePaiementRequest {
  montant: number;
  mode: CreatePaiementRequestMode;
}

export interface PaiementDetail {
  id: number;
  colisCode: string;
  montant: number;
  datePaiement: string;
  mode: string;
  montantDu: number;
  soldeApres: number;
}

export interface DetteDetail {
  colisProprietaireId: number;
  clientId: number;
  clientNom: string;
  clientTelephone: string;
  colisCode: string;
  arrivageDate: string;
  montantDu: number;
  montantPaye: number;
  solde: number;
}

export type DashboardDataToday = {
  nbColis: number;
  poidsTotal: number;
  montantTotal: number;
  montantEncaisse: number;
  dettesTotal: number;
};

export type DashboardDataEvolutionColisItem = {
  date: string;
  nbColis: number;
  montant: number;
};

export type DashboardDataTopClientsItem = {
  clientId: number;
  clientNom: string;
  totalDu: number;
  solde: number;
};

export interface DashboardData {
  today: DashboardDataToday;
  evolutionColis: DashboardDataEvolutionColisItem[];
  topClients: DashboardDataTopClientsItem[];
  totalDettesGlobal: number;
}

export interface BilanJournalier {
  date: string;
  nbColisIndividuels: number;
  nbColisMixtes: number;
  poidsIndividuels: number;
  poidsMixtes: number;
  montantTotal: number;
  montantEncaisse: number;
  montantDu: number;
  dettes: DetteDetail[];
  colis: ColisDetail[];
}

export type ListClientsParams = {
  search?: string;
};

export type ListArrivagesParams = {
  dateDebut?: string;
  dateFin?: string;
};

export type ListColisParams = {
  arrivageId?: number;
  clientId?: number;
  type?: ListColisType;
  dateDebut?: string;
  dateFin?: string;
};

export type ListColisType = (typeof ListColisType)[keyof typeof ListColisType];

export const ListColisType = {
  individuel: "individuel",
  mixte: "mixte",
} as const;

export type ListDettesParams = {
  clientId?: number;
  dateDebut?: string;
  dateFin?: string;
};

export type GetDashboardParams = {
  date?: string;
};

export type GetBilanJournalierParams = {
  date: string;
};