# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Import/Export management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clients/{id}/historique-paiements:
    get:
      operationId: getClientPaiements
      tags: [clients]
      summary: Get payment history for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaiementDetail"
  /arrivages:
    get:
      operationId: listArrivages
      tags: [arrivages]
      summary: List all shipments
      parameters:
        - name: dateDebut
          in: query
          schema:
            type: string
            format: date
        - name: dateFin
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of shipments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Arrivage"
    post:
      operationId: createArrivage
      tags: [arrivages]
      summary: Create a new shipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArrivageRequest"
      responses:
        "201":
          description: Shipment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Arrivage"
  /arrivages/{id}:
    get:
      operationId: getArrivage
      tags: [arrivages]
      summary: Get shipment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArrivageDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateArrivage
      tags: [arrivages]
      summary: Update a shipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArrivageRequest"
      responses:
        "200":
          description: Updated shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Arrivage"
    delete:
      operationId: deleteArrivage
      tags: [arrivages]
      summary: Delete a shipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Client:
      type: object
      properties:
        id:
          type: integer
        nom:
          type: string
        telephone:
          type: string
        adresse:
          type: string
          nullable: true
        dateCreation:
          type: string
          format: date-time
        totalDu:
          type: number
        totalPaye:
          type: number
        solde:
          type: number
      required:
        - id
        - nom
        - telephone
        - dateCreation
        - totalDu
        - totalPaye
        - solde
    CreateClientRequest:
      type: object
      properties:
        nom:
          type: string
        telephone:
          type: string
        adresse:
          type: string
          nullable: true
      required:
        - nom
        - telephone
    Arrivage:
      type: object
      properties:
        id:
          type: integer
        dateArrivee:
          type: string
          format: date
        nbColisTotal:
          type: integer
        poidsTotal:
          type: number
        coutTotal:
          type: number
        dateCreation:
          type: string
          format: date-time
      required:
        - id
        - dateArrivee
        - nbColisTotal
        - poidsTotal
        - coutTotal
        - dateCreation
    ArrivageDetail:
      allOf:
        - $ref: "#/components/schemas/Arrivage"
        - type: object
          properties:
            colis:
              type: array
              items:
                $ref: "#/components/schemas/ColisDetail"
          required:
            - colis
    CreateArrivageRequest:
      type: object
      properties:
        dateArrivee:
          type: string
          format: date
        nbColisTotal:
          type: integer
        poidsTotal:
          type: number
        coutTotal:
          type: number
      required:
        - dateArrivee
        - nbColisTotal
        - poidsTotal
        - coutTotal
    Colis:
      type: object
      properties:
        id:
          type: integer
        arrivageId:
          type: integer
        codeColisReel:
          type: string
        codeColisComplet:
          type: string
        type:
          type: string
          enum: [individuel, mixte]
        poids:
          type: number
        montant:
          type: number
        dateCreation:
          type: string
          format: date-time
      required:
        - id
        - arrivageId
        - codeColisReel
        - codeColisComplet
        - type
        - poids
        - montant
        - dateCreation
    ColisProprietaire:
      type: object
      properties:
        id:
          type: integer
        colisId:
          type: integer
        clientId:
          type: integer
        clientNom:
          type: string
        clientTelephone:
          type: string
        poids:
          type: number
        montantDu:
          type: number
        montantPaye:
          type: number
        solde:
          type: number
      required:
        - id
        - colisId
        - clientId
        - clientNom
        - clientTelephone
        - poids
        - montantDu
        - montantPaye
        - solde
    ColisDetail:
      allOf:
        - $ref: "#/components/schemas/Colis"
        - type: object
          properties:
            proprietaires:
              type: array
              items:
                $ref: "#/components/schemas/ColisProprietaire"
            arrivageDate:
              type: string
              format: date
          required:
            - proprietaires
            - arrivageDate
...
    PaiementDetail:
      type: object
      properties:
        id:
          type: integer
        colisCode:
          type: string
        montant:
          type: number
        datePaiement:
          type: string
          format: date-time
        mode:
          type: string
        montantDu:
          type: number
        soldeApres:
          type: number
      required:
        - id
        - colisCode
        - montant
        - datePaiement
        - mode
        - montantDu
        - soldeApres
...