# lib/api-zod/src/generated/types/createArrivageRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Import/Export management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateArrivageRequest {
  dateArrivee: Date;
  nbColisTotal: number;
  poidsTotal: number;
  coutTotal: number;
}