# lib/api-zod/src/generated/types/paiementDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Import/Export management API
 * OpenAPI spec version: 0.1.0
 */

export interface PaiementDetail {
  id: number;
  colisCode: string;
  montant: number;
  datePaiement: Date;
  mode: string;
  montantDu: number;
  soldeApres: number;
}